Separate parser code into its own file
Parser code will live in its own file and not in
the cli file.
